fastapi-ecommerce/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
pydantic[email]
pymysql
streamlit
pandas
python-dotenv
alembic